+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#29] Remove implementation of the remote http dogu registry and instead, reuse the implementation from the cesapp-lib. 
 
 ## [v0.8.0] - 2022-06-08
 ### Added